fix: Rename hook of pagination
